% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ActivityDateType}
\alias{ActivityDateType}
\title{ActivityDateType}
\format{
A data frame with 4 rows and 6 variables:
\describe{
  \item{\code{code}}{double COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{description}}{character COLUMN_DESCRIPTION}
  \item{\code{category}}{logical COLUMN_DESCRIPTION}
  \item{\code{url}}{logical COLUMN_DESCRIPTION}
  \item{\code{status}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
ActivityDateType
}
\description{
Types of date for activities.
There are many different business models and dates that can be used to describe
 the start and end of activities. It is recommended that each publisher
 adopts their own consistent approach that provides users with a meaningful
 indication of the lifespan of an activity.
}
\keyword{datasets}
