% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AidType}
\alias{AidType}
\title{AidType}
\format{
A data frame with 18 rows and 6 variables:
\describe{
  \item{\code{code}}{character COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{description}}{character COLUMN_DESCRIPTION}
  \item{\code{category}}{character COLUMN_DESCRIPTION}
  \item{\code{url}}{logical COLUMN_DESCRIPTION}
  \item{\code{status}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
AidType
}
\description{
See: http://www.oecd.org/dac/stats/dacandcrscodelists.htm
}
\examples{
{
head(AidType,10)
}
}
\keyword{datasets}
