% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AidTypeCategory}
\alias{AidTypeCategory}
\title{AidTypeCategory}
\format{
A data frame with 8 rows and 6 variables:
\describe{
  \item{\code{code}}{character COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{description}}{character COLUMN_DESCRIPTION}
  \item{\code{category}}{logical COLUMN_DESCRIPTION}
  \item{\code{url}}{logical COLUMN_DESCRIPTION}
  \item{\code{status}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
AidTypeCategory
}
\description{
This codelists exists to group the Aid Type codelist into
categories. It is not used as a codelist in its own right.
External URL: http://www.oecd.org/dac/stats/dacandcrscodelists.htm
}
\keyword{datasets}
