% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BudgetIdentifier}
\alias{BudgetIdentifier}
\title{BudgetIdentifier}
\format{
A data frame with 152 rows and 6 variables:
\describe{
  \item{\code{code}}{character COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{description}}{logical COLUMN_DESCRIPTION}
  \item{\code{category}}{double COLUMN_DESCRIPTION}
  \item{\code{url}}{logical COLUMN_DESCRIPTION}
  \item{\code{status}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
BudgetIdentifier
}
\description{
IATI Functional and Administrative Common Code :
One of several possible Budget Identifier Vocabularies.
As of version 2.03 this codelist has been deprecated.
  This is a Non-Core codelist.
}
\keyword{datasets}
