% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CashandVoucherModalities}
\alias{CashandVoucherModalities}
\title{CashandVoucherModalities}
\format{
A data frame with 2 rows and 6 variables:
\describe{
  \item{\code{code}}{double COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{description}}{character COLUMN_DESCRIPTION}
  \item{\code{category}}{logical COLUMN_DESCRIPTION}
  \item{\code{url}}{logical COLUMN_DESCRIPTION}
  \item{\code{status}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
CashandVoucherModalities
}
\description{
This codelist has been created by IATI following agreements
and recommendations of the Tracking Cash and Voucher Assistance (CVA)
Working Group. Definitions of the codes have been aligned with the
CaLP Glossary:http://www.cashlearning.org/resources/glossary.
  This is a Non-Core codelist.
}
\examples{
{
head(CashandVoucherModalities,10)
}
}
\keyword{datasets}
