% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Country}
\alias{Country}
\title{Country}
\format{
A data frame with 251 rows and 6 variables:
\describe{
  \item{\code{code}}{character COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{description}}{logical COLUMN_DESCRIPTION}
  \item{\code{category}}{logical COLUMN_DESCRIPTION}
  \item{\code{url}}{logical COLUMN_DESCRIPTION}
  \item{\code{status}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
Country
}
\description{
The Country codelist is generated from the ISO 3166-1 part
 of the ISO 3166 standard. The standard makes allowance, alongside
  the officially assigned codes, for code elements to be expanded by using
  either reserved codes or user-assigned codes.
  IATI currently defines additional codes in the XA -XZ range.
    External URL: http://www.iso.org/iso/home/standards/country_codes.htm
}
\examples{
{
head(Country,10)
}
}
\keyword{datasets}
