% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FlowType}
\alias{FlowType}
\title{FlowType}
\format{
A data frame with 10 rows and 6 variables:
\describe{
  \item{\code{code}}{double COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{description}}{character COLUMN_DESCRIPTION}
  \item{\code{category}}{logical COLUMN_DESCRIPTION}
  \item{\code{url}}{logical COLUMN_DESCRIPTION}
  \item{\code{status}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
FlowType
}
\description{
DAC/CRS distinction between ODA (official development assistance)
 and other types of resource flow.
   External URL: http://www.oecd.org/dac/stats/dacandcrscodelists.htm
}
\examples{
{
head(FlowType,10)
}
}
\keyword{datasets}
