% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIATI.R
\name{get_IATI}
\alias{get_IATI}
\title{get_IATI}
\usage{
get_IATI(url, query, value)
}
\arguments{
\item{url}{URL for websites containing IATI-compliant datasets.}

\item{query}{Search/query parameter. For IATI activites, this can be
`iati_identifier`, `title`, `description`, `recipient_country`,
`recipient_region`, `recipient_org`, `sector`, `document_link`,
`participating_org`.}

\item{value}{Value to pass to specified `query`.}
}
\value{
Parsed content on activities from IATI-compliant datasets.
}
\description{
Function to get activities endpoint from IATI-compliant datasets.
}
\examples{
get_activities(url = "https://devtracker.dfid.gov.uk",
               query = "iati_identifier",
               value = "075004")

}
