% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HumanitarianScopeType}
\alias{HumanitarianScopeType}
\title{HumanitarianScopeType}
\format{
A data frame with 2 rows and 6 variables:
\describe{
  \item{\code{code}}{double COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{description}}{logical COLUMN_DESCRIPTION}
  \item{\code{category}}{logical COLUMN_DESCRIPTION}
  \item{\code{url}}{logical COLUMN_DESCRIPTION}
  \item{\code{status}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
HumanitarianScopeType
}
\description{
The Humanitarian Scope Type codelist defines codes for types of
humanitarian events and actions.
}
\examples{
{
head(HumanitarianScopeType,10)
}
}
\keyword{datasets}
