% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HumanitarianScopeVocabulary}
\alias{HumanitarianScopeVocabulary}
\title{HumanitarianScopeVocabulary}
\format{
A data frame with 3 rows and 6 variables:
\describe{
  \item{\code{code}}{character COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{description}}{logical COLUMN_DESCRIPTION}
  \item{\code{category}}{logical COLUMN_DESCRIPTION}
  \item{\code{url}}{character COLUMN_DESCRIPTION}
  \item{\code{status}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
HumanitarianScopeVocabulary
}
\description{
The Humanitarian Scope Vocabulary codelist defines a range of
external codelists which themselves provide codes and descriptions for
humanitarian events and actions.
}
\keyword{datasets}
