% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HumCluster}
\alias{HumCluster}
\title{HumCluster}
\format{
A data frame with 16 rows and 9 variables:
\describe{
  \item{\code{HRinfo ID}}{character COLUMN_DESCRIPTION}
  \item{\code{Preferred Term}}{character COLUMN_DESCRIPTION}
  \item{\code{Preferred Term (fr)}}{character COLUMN_DESCRIPTION}
  \item{\code{ACRONYM}}{character COLUMN_DESCRIPTION}
  \item{\code{Group Type}}{character COLUMN_DESCRIPTION}
  \item{\code{Homepage}}{character COLUMN_DESCRIPTION}
  \item{\code{Date Created (Post Date)}}{character COLUMN_DESCRIPTION}
  \item{\code{Updated date}}{character COLUMN_DESCRIPTION}
  \item{\code{Notes}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
HumCluster
}
\description{
Global Coordination Groups - Cluster & Area of Interest
Official Names and Codes (Beta)
Contains IASC designated Global Clusters and Areas of Responsibility.
Global Clusters and Areas of Responsibility (AoR) were defined to enable more
 predictable leadership in situations of humanitarian emergency.
 The content includes the Preferred Term to be used for each Cluster and AoR,
  the Acronym (3-letter cluster code), the Group to which the term belongs
  (defining whether it is a Global Cluster or AoR) and the website URL.
   Both the Google Spreadsheet and CSV versions contain HXL hashtags.
     For more information, please visit http://vocabulary.unocha.org/
}
\keyword{datasets}
