% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IATIOrganisationIdentifier}
\alias{IATIOrganisationIdentifier}
\title{IATIOrganisationIdentifier}
\format{
A data frame with 39 rows and 6 variables:
\describe{
  \item{\code{code}}{character COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{description}}{character COLUMN_DESCRIPTION}
  \item{\code{category}}{logical COLUMN_DESCRIPTION}
  \item{\code{url}}{character COLUMN_DESCRIPTION}
  \item{\code{status}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
IATIOrganisationIdentifier
}
\description{
This is a list of organisation identifiers that is maintained by
 the IATI Secretariat. The prefix for organisations on this list is XI-IATI
   If a bona fide organisation is not registered with any recognised or
   appropriate registration agency
   (http://iatistandard.org/202/codelists/OrganisationRegistrationAgency/)
   they should contact the IATI Technical Team who will exceptionally allocate
    an organisation identifier using the XI-IATI prefix.

    While some of these identifiers have been derived from DAC codes, this
    ‘meaning’ is not carried forward. i.e. IATI generated identifiers have no
     intrinsic meaning.
   For general guidance about constructing Organisation Identifiers, please see
    http://iatistandard.org/organisation-identifiers/
}
\examples{
{
head(IATIOrganisationIdentifier,10)
}
}
\keyword{datasets}
