% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IatiTidy-package.R
\docType{package}
\name{IatiTidy-package}
\alias{IatiTidy}
\alias{IatiTidy-package}
\title{IatiTidy: Tidy Data from IATI Data Store & FTS API}
\description{
Wrapper for IATI DataStore: https://iatidatastore.iatistandard.org/querybuilder 
 and FTS API https://api.hpc.tools/docs/v1/
}
\seealso{
Useful links:
\itemize{
  \item \url{https://unhcr-americas.github.io/IatiTidy/docs/}
  \item Report bugs at \url{https://github.com/unhcr-americas/IatiTidy/issues}
}

}
\keyword{internal}
