% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implementers_split.R
\name{implementers_split}
\alias{implementers_split}
\title{implementers_split}
\usage{
implementers_split(row)
}
\arguments{
\item{row}{lines in the query extract}
}
\value{
Parsed content on activities from IATI-compliant datasets.
}
\description{
Function to split activities by implementers.
Thanks to https://github.com/akmiller01/IATI-Covid
}
