% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IndicatorVocabulary}
\alias{IndicatorVocabulary}
\title{IndicatorVocabulary}
\format{
A data frame with 10 rows and 6 variables:
\describe{
  \item{\code{code}}{double COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{description}}{logical COLUMN_DESCRIPTION}
  \item{\code{category}}{logical COLUMN_DESCRIPTION}
  \item{\code{url}}{character COLUMN_DESCRIPTION}
  \item{\code{status}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
IndicatorVocabulary
}
\description{
The Indicator Vocabulary codelist defines a range of external
codelists which themselves provide codes and descriptions for indicators,
for example to specify results.
}
\examples{
{
head(IndicatorVocabulary,10)
}
}
\keyword{datasets}
