% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OrganisationIdentifier}
\alias{OrganisationIdentifier}
\title{OrganisationIdentifier}
\format{
A data frame with 419 rows and 6 variables:
\describe{
  \item{\code{code}}{character COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{description}}{logical COLUMN_DESCRIPTION}
  \item{\code{category}}{logical COLUMN_DESCRIPTION}
  \item{\code{url}}{logical COLUMN_DESCRIPTION}
  \item{\code{status}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
OrganisationIdentifier
}
\description{
As of 1.04 this list is no longer maintained.
 http://support.iatistandard.org/entries/28497976-Retire-the-Organisation-Identifier-codelist#view-post-25368673
   For general guidance about constructing Organisation Identifiers,
   please see http://iatistandard.org/organisation-identifiers/
}
\examples{
{
head(OrganisationIdentifier,10)
}
}
\keyword{datasets}
