% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OrganisationRegistrationAgency}
\alias{OrganisationRegistrationAgency}
\title{OrganisationRegistrationAgency}
\format{
A data frame with 197 rows and 7 variables:
\describe{
  \item{\code{code}}{character COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{description}}{character COLUMN_DESCRIPTION}
  \item{\code{category}}{character COLUMN_DESCRIPTION}
  \item{\code{url}}{character COLUMN_DESCRIPTION}
  \item{\code{status}}{character COLUMN_DESCRIPTION}
  \item{\code{public-database}}{logical COLUMN_DESCRIPTION}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
OrganisationRegistrationAgency
}
\description{
The values from this codelist are used to identify the particular
 list that an organisation identifier was drawn from. The codelist provides a
 register of known identifier lists, including national company registers,
 NGO directories and international and multilateral organisation lists -
 along with guidance and online resources to help locate the identifiers
 assigned to a specific organisation.
 As of 17 July 2017 this list is maintained by the org-id.guide project.
 Data publishers can now search for and locate the relevant list for a
 particular organisation identifier using the org-id.guide website.

 The full register of identifier sources is also available to download
 in XML, JSON and CSV formats.

 IATI periodically replicates the codelist of identifier sources
 from org-id.guide, to assist those accessing IATI documentation.
 However, it is advised that the most up-to-date source is the
 org-id.guide project.

 If org-id.guide does not contain an entry for the kind of
 organisation you need to identify, you can make a request a
 new list is included in the register following the guidance
}
\keyword{datasets}
