% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PolicyMarker}
\alias{PolicyMarker}
\title{PolicyMarker}
\format{
A data frame with 12 rows and 6 variables:
\describe{
  \item{\code{code}}{double COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{description}}{logical COLUMN_DESCRIPTION}
  \item{\code{category}}{logical COLUMN_DESCRIPTION}
  \item{\code{url}}{logical COLUMN_DESCRIPTION}
  \item{\code{status}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
PolicyMarker
}
\description{
The Policy Marker codelist is derived from the policy markers
 declared by the WP-STAT. The codes themselves are created by IATI.
   External URL: http://www.oecd.org/dac/stats/dacandcrscodelists.htm
}
\keyword{datasets}
