% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sector}
\alias{Sector}
\title{Sector}
\format{
A data frame with 323 rows and 6 variables:
\describe{
  \item{\code{code}}{double COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{description}}{character COLUMN_DESCRIPTION}
  \item{\code{category}}{double COLUMN_DESCRIPTION}
  \item{\code{url}}{logical COLUMN_DESCRIPTION}
  \item{\code{status}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
Sector
}
\description{
DAC 5 Digit Sector
  External URL: http://www.oecd.org/dac/stats/dacandcrscodelists.htm
}
\examples{
{
head(Sector,10)
}
}
\keyword{datasets}
