% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SectorCategory}
\alias{SectorCategory}
\title{SectorCategory}
\format{
A data frame with 44 rows and 6 variables:
\describe{
  \item{\code{code}}{double COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{description}}{character COLUMN_DESCRIPTION}
  \item{\code{category}}{logical COLUMN_DESCRIPTION}
  \item{\code{url}}{logical COLUMN_DESCRIPTION}
  \item{\code{status}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
SectorCategory
}
\description{
DAC 3 Digit Sector
  External URL: http://www.oecd.org/dac/stats/dacandcrscodelists.htm
}
\keyword{datasets}
