% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SectorVocabulary}
\alias{SectorVocabulary}
\title{SectorVocabulary}
\format{
A data frame with 13 rows and 6 variables:
\describe{
  \item{\code{code}}{double COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{description}}{character COLUMN_DESCRIPTION}
  \item{\code{category}}{logical COLUMN_DESCRIPTION}
  \item{\code{url}}{character COLUMN_DESCRIPTION}
  \item{\code{status}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
SectorVocabulary
}
\description{
This is a Non-Core codelist.
}
\keyword{datasets}
