% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_vocabulary.R
\name{single_vocabulary}
\alias{single_vocabulary}
\title{single_vocabulary}
\usage{
single_vocabulary(row)
}
\arguments{
\item{row}{lines in the query extract}
}
\value{
Parsed content on activities from IATI-compliant datasets.
}
\description{
Function to split sectors by activities.
Thanks to https://github.com/akmiller01/IATI-Covid
}
