% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_countries.R, R/split_implementers.R
\name{split_implementers}
\alias{split_implementers}
\title{split_implementers}
\usage{
split_implementers(data)

split_implementers(data)
}
\arguments{
\item{row}{lines in the query extract}
}
\value{
Parsed content on activities from IATI-compliant datasets.

Parsed content on activities from IATI-compliant datasets.
}
\description{
Function to split activities by implementers.
Thanks to https://github.com/akmiller01/IATI-Covid

Function to split activities by implementers.
Thanks to https://github.com/akmiller01/IATI-Covid
}
