% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_sector.R
\name{split_sector}
\alias{split_sector}
\title{split_sector}
\usage{
split_sector(data)
}
\arguments{
\item{data}{query extract}
}
\value{
Parsed content on activities from IATI-compliant datasets.
}
\description{
Function to split sectors by activities.
Thanks to https://github.com/akmiller01/IATI-Covid
}
