% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_transaction.R
\name{split_transaction}
\alias{split_transaction}
\title{split_transaction}
\usage{
split_transaction(row)
}
\arguments{
\item{row}{lines in the query extract}
}
\value{
Parsed content on activities from IATI-compliant datasets.
}
\description{
Function to split activities by implementers.
Thanks to https://github.com/akmiller01/IATI-Covid
}
