% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UNSDGGoals}
\alias{UNSDGGoals}
\title{UNSDGGoals}
\format{
A data frame with 17 rows and 6 variables:
\describe{
  \item{\code{code}}{double COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{description}}{logical COLUMN_DESCRIPTION}
  \item{\code{category}}{logical COLUMN_DESCRIPTION}
  \item{\code{url}}{logical COLUMN_DESCRIPTION}
  \item{\code{status}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
UNSDGGoals
}
\description{
A value from the top-level list of UN sustainable development
goals (SDGs) (e.g. ‘1’)
  External URL: https://sustainabledevelopment.un.org/?menu=1300
}
\keyword{datasets}
