% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UNSDGTargets}
\alias{UNSDGTargets}
\title{UNSDGTargets}
\format{
A data frame with 169 rows and 6 variables:
\describe{
  \item{\code{code}}{character COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{description}}{logical COLUMN_DESCRIPTION}
  \item{\code{category}}{logical COLUMN_DESCRIPTION}
  \item{\code{url}}{logical COLUMN_DESCRIPTION}
  \item{\code{status}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://iatistandard.org/en/iati-standard/203/codelists/}
}
\usage{
UNSDGTargets
}
\description{
A value from the second-level list of UN sustainable development
goals (SDGs) (e.g. ‘1.1’)
  External URL: https://unstats.un.org/sdgs/indicators/indicators-list/
}
\keyword{datasets}
