% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FTS.R
\name{get_FTS}
\alias{get_FTS}
\title{get_FTS}
\usage{
get_FTS(
  url = "https://api.hpc.tools",
  ext = "/v1/public/fts/flow?",
  year = "year=2019,2020"
)
}
\arguments{
\item{url}{URL for FTS API}

\item{ext}{API endpoint}

\item{year}{year to include}
}
\value{
flows_frame data frame with flows
}
\description{
Function to get flows from FTS.
}
\examples{
\dontrun{
data <- get_FTS(url = "https://api.hpc.tools",
                             ext = "/v1/public/fts/flow?",
                             year = "year=2019,2020")
}

}
